+++ Changes.md
@@ -5,6 +5,7 @@
 HEAD
 ----------
 
+- Job iteration now exposes `current_object` for easy access within the `around_iteration` callback [#6774]
 - Fix JS race condition which could skip confirmation dialogs when Live Polling [#6768]
 - Fix edge case which could lose CurrentAttributes [#6767]